requirements.txt
Flask>=2.0
polars>=0.18.0
numpy
requests
google-generativeai>=0.4.0
